+++ imod/mf6/cnc.py
@@ -1,5 +1,6 @@
 import numpy as np
 
+from imod.logging.logging_decorators import init_log_decorator
 from imod.mf6.boundary_condition import BoundaryCondition
 from imod.mf6.interfaces.iregridpackage import IRegridPackage
 from imod.mf6.validation import BOUNDARY_DIMS_SCHEMA
@@ -66,6 +67,7 @@ class ConstantConcentration(BoundaryCondition, IRegridPackage):
         ]
     }
 
+    @init_log_decorator()
     def __init__(
         self,
         concentration,