+++ imod/mf6/model.py
@@ -18,6 +18,7 @@ import xugrid as xu
 from jinja2 import Template
 
 import imod
+from imod.logging.logging_decorators import standard_log_decorator
 from imod.mf6.interfaces.imodel import IModel
 from imod.mf6.package import Package
 from imod.mf6.statusinfo import NestedStatusInfo, StatusInfo, StatusInfoBase
@@ -185,7 +186,8 @@ class Modflow6Model( collections.UserDict, IModel, abc.ABC):
 
         k = npf["k"]
         return k
-
+    
+    @standard_log_decorator()
     def validate(self, model_name: str = "") -> StatusInfoBase:
         try:
             diskey = self._get_diskey()
@@ -224,7 +226,8 @@ class Modflow6Model( collections.UserDict, IModel, abc.ABC):
                 model_status_info.add(pkg_errors_to_status_info(pkg_name, pkg_errors))
 
         return model_status_info
-
+    
+    @standard_log_decorator()
     def write(
         self, modelname, globaltimes, validate: bool, write_context: WriteContext
     ) -> StatusInfoBase:
@@ -290,6 +293,7 @@ class Modflow6Model( collections.UserDict, IModel, abc.ABC):
 
         return NestedStatusInfo(modelname)
 
+    @standard_log_decorator()
     def dump(
         self, directory, modelname, validate: bool = True, mdal_compliant: bool = False
     ):