+++ imod/mf6/mst.py
@@ -2,6 +2,7 @@ from typing import Optional, Tuple
 
 import numpy as np
 
+from imod.logging.logging_decorators import init_log_decorator
 from imod.mf6.interfaces.iregridpackage import IRegridPackage
 from imod.mf6.package import Package
 from imod.mf6.utilities.regrid import RegridderType
@@ -111,7 +112,7 @@ class MobileStorageTransfer(Package, IRegridPackage):
         "distcoef": (RegridderType.OVERLAP, "mean"),
         "sp2": (RegridderType.OVERLAP, "mean"),        
     }
-
+    @init_log_decorator()
     def __init__(
         self,
         porosity,