+++ imod/mf6/rch.py
@@ -2,6 +2,7 @@ from typing import Optional, Tuple
 
 import numpy as np
 
+from imod.logging.logging_decorators import init_log_decorator
 from imod.mf6.boundary_condition import BoundaryCondition
 from imod.mf6.interfaces.iregridpackage import IRegridPackage
 from imod.mf6.utilities.regrid import RegridderType
@@ -111,6 +112,7 @@ class Recharge(BoundaryCondition, IRegridPackage):
         "concentration": (RegridderType.OVERLAP, "mean"),
     }
 
+    @init_log_decorator()
     def __init__(
         self,
         rate,