+++ docs/api/changelog.rst
@@ -46,6 +46,9 @@ Added
   :meth:`imod.mf6.Modflow6Simulation.open_transport_budget` support opening
   split multi-species simulations.
   :meth:`imod.mf6.Modflow6Simulation.regrid_like` can now regrid simulations that have 1 or more transport models.
+- added logging to various initialization methods, write methods and dump methods. `See the
+  documentation <https://deltares.github.io/imod-python/api/generated/logging/imod.logging.html>`_
+  how to activate logging.
 
 Changed
 ~~~~~~~